@@ -1,1 +1,1 @@
-ruby-2.0.0
+ruby-2.4
