@@ -1,7 +1,6 @@
 source 'http://rubygems.org'
 
-gem 'railties', '3.2.22.4'
-gem 'rails', '3.2.22.4'
+gem 'rails', '4.2.8'
 
 gem 'json', '1.8.6'
 gem 'haml', '4.0.5'
@@ -12,9 +11,10 @@ gem 'rdiscount', '1.6.8'
 gem 'rack-no-www', '0.0.2'
 gem 'httparty', '0.13.1'
 
+gem 'rack-cache', '1.7.0'
+
 group :test, :development do
-  gem 'sqlite3', '1.3.10'
-  gem 'rspec-rails', '3.0.2'
+  gem 'rspec-rails', '3.5.2'
   gem 'ZenTest', '4.11.0'
   gem 'spork', '0.9.2'
   gem 'autotest', '4.4.6'
@@ -23,15 +23,23 @@ group :test, :development do
   gem 'webrat', '0.7.3'
   gem 'database_cleaner', '1.3.0'
   gem 'test-unit', '3.1.4'
-  gem 'cucumber-rails', '1.4.1'
+
+  gem 'pry'
+  gem 'pry-byebug'
+  gem 'activerecord-deprecated_finders', require: 'active_record/deprecated_finders'
+  gem 'sqlite3'
+end
+
+group :test do
+  gem 'cucumber-rails', '1.4.5'
 end
 
 # Gems used only for assets and not required
 # in production environments by default.
 group :assets do
-  gem 'sass-rails', '3.2.6'
-  gem 'coffee-rails', '3.2.2'
-  gem 'uglifier', '2.5.3'
+  gem 'sass-rails', '5.0.6'
+  gem 'coffee-rails', '4.2.1'
+  gem 'uglifier', '3.2.0'
 end
 
-gem 'jquery-rails', '3.1.1'
+gem 'jquery-rails'
